+++ zarrs/src/array/codec/bytes_to_bytes/bz2.rs
@@ -38,7 +38,6 @@ use std::sync::Arc;
 
 use crate::{
     array::codec::{Codec, CodecPlugin},
-    config::global_config,
     metadata::codec::bz2,
     metadata::v3::MetadataV3,
     plugin::{PluginCreateError, PluginMetadataInvalidError},
@@ -54,14 +53,11 @@ pub use bz2::IDENTIFIER;
 
 // Register the codec.
 inventory::submit! {
-    CodecPlugin::new(IDENTIFIER, is_name_bz2, create_codec_bz2)
+    CodecPlugin::new(IDENTIFIER, is_identifier_bz2, create_codec_bz2)
 }
 
-fn is_name_bz2(name: &str) -> bool {
-    global_config()
-        .codec_map()
-        .get(IDENTIFIER)
-        .is_some_and(|map| map.contains(name))
+fn is_identifier_bz2(identifier: &str) -> bool {
+    identifier == IDENTIFIER
 }
 
 pub(crate) fn create_codec_bz2(metadata: &MetadataV3) -> Result<Codec, PluginCreateError> {