+++ zarrs_metadata/CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for a `must_understand` field to `MetadataV3` (ZEP0009)
   - Extensions can now be parsed in more than just the additional fields of array/group metadata (e.g. codecs)
   - Automatically skip unsupported codecs/storage transformers with `"must_understand": false`
-- Add `CodecMap` and `CodecName` for codec `nam` overriding and aliasing
+- Add `ExtensionMaps[Codec]` and `Extension{Name,Alias}Map` for extension `name` overriding and aliasing
+  - This API will also be introduced for data types in the future.
 - Implement `From<T> for MetadataConfiguration` for all codec configuration enums
 - Implement `Copy` for `ZstdCompressionLevel`
 - Add `zfpy` codec metadata (unmerged from `zfp`)
@@ -23,8 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: Refactor `FillValueMetadataV3` to support arbitrary fill value metadata (for ZEP0009)
 - **Breaking**: Rename `DataTypeMetadataV3::Unknown` variant to `Extension`
 - **Breaking**: Mark versioned codec metadata as non-exhaustive
-- **Breaking**: `{array,codec}_metadata_v2_to_v3` have an additional `CodecMap` parameter
-  - `zarrs` has a default codec map accessible via `zarrs::config::global_config().codec_map()`
+- **Breaking**: `{array,codec}_metadata_v2_to_v3` have an additional `ExtensionMapsCodec` parameter
+  - `zarrs` has a default codec map accessible via `zarrs::config::global_config().codec_maps()`
 - **Breaking**: Remove `write_header` from `zfp` codec configuration
 - Bump `half` to 2.3.1
 - Bump `thiserror` to 2.0.12